use int32 instead of int where needed in leaky bucket and rate limiter for grpc

+++ flow/barito_producer_service.go
@@ -160,7 +160,7 @@ func (s *baritoProducerService) ServeHTTP(rw http.ResponseWriter, req *http.Requ
 		// add suffix
 		topic = timberCollection.Context.KafkaTopic + s.topicSuffix
 
-		maxTokenIfNotExist := timberCollection.Context.AppMaxTPS
+		maxTokenIfNotExist := int32(timberCollection.Context.AppMaxTPS)
 		if s.limiter.IsHitLimit(topic, len(timberCollection.Items), maxTokenIfNotExist) {
 			onLimitExceeded(rw)
 			return
@@ -189,7 +189,7 @@ func (s *baritoProducerService) ServeHTTP(rw http.ResponseWriter, req *http.Requ
 		// add suffix
 		topic = timber.Context().KafkaTopic + s.topicSuffix
 
-		maxTokenIfNotExist := timber.Context().AppMaxTPS
+		maxTokenIfNotExist := int32(timber.Context().AppMaxTPS)
 		if s.limiter.IsHitLimit(topic, 1, maxTokenIfNotExist) {
 			onLimitExceeded(rw)
 			return
